api/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
onnxruntime>=1.15.0
numpy>=1.24.0
prometheus-client>=0.17.0
python-multipart>=0.0.6
pytest>=7.3.1
httpx>=0.24.1
requests>=2.31.0
google-generativeai>=0.1.0
python-dotenv>=1.0.0